+++ R/graph_proxies.R
@@ -289,7 +289,7 @@ sg_add_edges_delay_p <- function(proxy, data, delay, ..., refresh = TRUE, cumsum
 	# build data
 	nodes <- .build_data(data, ...) %>%
 		dplyr::bind_cols(delay_table) %>% # bind delay
-	.check_ids() %>%
+	  .check_ids() %>%
 		.check_x_y() %>%
 		.as_list()
 
@@ -361,4 +361,99 @@ sg_drop_edges_p <- function(proxy, data, ids, refresh = TRUE, rate = "once") {
 	proxy$session$sendCustomMessage("sg_drop_edges_p", message)
 
 	return(proxy)
+}
+
+#' Drop nodes or edges with a delay
+#' 
+#' Proxies to dynamically drop multiple nodes or edges to an already existing graph with a *delay* between each removal.
+#' 
+#' @param proxy An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.
+#' @param data A \code{data.frame} of _one_ node or edge.
+#' @param ... any column.
+#' @param refresh Whether to refresh the graph after node is dropped, required to take effect, if you are running force the algorithm is killed and restarted at every iteration.
+#' @param delay Column name containing delay in milliseconds.
+#' @param cumsum Whether to compute the cumulative sum of the delay.
+#' 
+#' @details The delay helps for build dynamic visualisations where nodes and edges do not disappear all at the same time.
+#' How the delay works depends on the \code{cumsum} parameter. if \code{TRUE} the function computes the cumulative sum
+#' of the delay to effectively drop each row one after the other: delay is thus applied at each row (number of seconds to wait
+#' before the row is dropped *since the previous row*). If \code{FALSE} this is the number of milliseconds to wait before the node or
+#' edge is added to the visualisation; \code{delay} is used as passed to the function.
+#'
+#' @examples
+#' \dontrun{
+#' demo("drop-nodes-delay", package = "sigmajs") # add nodes with a delay
+#' demo("drop-edges-delay", package = "sigmajs") # add edges with a delay
+#' demo("drop-delay", package = "sigmajs") # add nodes and edges with a delay
+#' }
+#'
+#' @note Have the parameters from your initial graph match that of the node you add, i.e.: if you pass \code{size} in your initial chart,
+#' make sure you also have it in your proxy.
+#' 
+#' @rdname drop_delay_p
+#' @export
+sg_drop_nodes_delay_p <- function(proxy, data, ids, delay, refresh = TRUE, rate = "once") {
+  
+  if (!"sigmajsProxy" %in% class(proxy))
+    stop("must pass sigmajsProxy object", call. = FALSE)
+  
+  if (missing(data))
+    stop("must pass data", call. = FALSE)
+  
+  delay_col <- eval(substitute(delay), data) # subset delay
+  if (isTRUE(cumsum))
+    delay_col <- cumsum(delay_col) # cumul for setTimeout
+  
+  ids <- eval(substitute(ids), data) # subset ids
+  
+  to_drop <- dplyr::tibble(
+    id = ids,
+    sigmajsdelay = delay_col
+  )
+  
+  to_drop <- dplyr::tibble(
+    id = ids,
+    sigmajsdelay = delay_col
+  ) %>% # bind delay
+    .as_list()
+  
+  message <- list(id = proxy$id, data = to_drop, refresh = refresh, rate = rate) # create message
+  
+  proxy$session$sendCustomMessage("sg_drop_nodes_delay_p", message)
+  
+  return(proxy)
+}
+
+#' @rdname adds_delay_p
+#' @export
+sg_drop_edges_delay_p <- function(proxy, data, ids, delay, refresh = TRUE) {
+  
+  if (!"sigmajsProxy" %in% class(proxy))
+    stop("must pass sigmajsProxy object", call. = FALSE)
+  
+  if (missing(data))
+    stop("must pass data", call. = FALSE)
+  
+  delay_col <- eval(substitute(delay), data) # subset delay
+  if (isTRUE(cumsum))
+    delay_col <- cumsum(delay_col) # cumul for setTimeout
+  
+  ids <- eval(substitute(ids), data) # subset ids
+  
+  to_drop <- dplyr::tibble(
+    id = ids,
+    sigmajsdelay = delay_col
+  )
+  
+  to_drop <- dplyr::tibble(
+    id = ids,
+    sigmajsdelay = delay_col
+  ) %>% # bind delay
+    .as_list()
+  
+  message <- list(id = proxy$id, data = to_drop, refresh = refresh) # create message
+  
+  proxy$session$sendCustomMessage("sg_drop_edges_delay_p", message)
+  
+  return(proxy)
 }