+++ NEWS.md
@@ -7,6 +7,8 @@ Functions added:
 * `sg_force_stop` - stops the forceAtlas2 layout after a given delay.
 * `sg_force_start` added, same as `sg_force`.
 * `sg_add_nodes`, `sg_add_edges` - Add nodes or edges to a static graph.
+* `drop_nodes_delay_p`, `drop_edges_delay_p` Proxy to drop nodes or edges with a delay.
+* `drop_nodes_delay`, `drop_edges_delay` Drop nodes or edges with a delay.
 
 # v0.0.1